+++ sdk/security/keyvault/azkeys/models.go
@@ -70,21 +70,18 @@ type ImportKeyOptions struct {
 // ListDeletedKeyPropertiesOptions contains the optional parameters for the Client.NewListDeletedKeyPropertiesPager
 // method.
 type ListDeletedKeyPropertiesOptions struct {
-	// Maximum number of results to return in a page. If not specified the service will return up to 25 results.
-	MaxResults *int32
+	// placeholder for future optional parameters
 }
 
 // ListKeyPropertiesOptions contains the optional parameters for the Client.NewListKeyPropertiesPager method.
 type ListKeyPropertiesOptions struct {
-	// Maximum number of results to return in a page. If not specified the service will return up to 25 results.
-	MaxResults *int32
+	// placeholder for future optional parameters
 }
 
 // ListKeyPropertiesVersionsOptions contains the optional parameters for the Client.NewListKeyPropertiesVersionsPager
 // method.
 type ListKeyPropertiesVersionsOptions struct {
-	// Maximum number of results to return in a page. If not specified the service will return up to 25 results.
-	MaxResults *int32
+	// placeholder for future optional parameters
 }
 
 // PurgeDeletedKeyOptions contains the optional parameters for the Client.PurgeDeletedKey method.